[refact] change shell main to command pattern
